requirements.txt
django
django-filter
djangorestframework
markdown
psycopg2
gnupg
pillow

# perso
ipython

# client
requests


# temp
pygments

